Moved delete fetch into fetchFunctions file.
